requirements.txt
streamlit==1.32.0
pandas==2.2.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
SQLAlchemy==2.0.27
google-generativeai==0.3.2
python-dotenv==1.0.1 